Added openspim.sh meant to live in /etc/profile.d/
Added openspim.sh meant to live in /etc/profile.d/

Add a script intended to live in msysGit's /etc/profile.d/

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

Fix tyop in the profile.d script to auto-build everything

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

cd to uManager in the profile.d script to auto-build everything

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

profile.d/ script: describe how to start Fiji

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

profile.d/ script: warn when there are differences to the committed version

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

profile.d/ script: install a start-openspim.sh script

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

profile.d/ script: if the script is checked out, run that instead

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

profile.d/ script: Always reinstate the desktop icon

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

Startup script: make sure jvisualvm is in the PATH

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

Switch the profile.d script to a Git for Windows one

There is no good reason anymore why we should require Fiji to be built
from scratch, or for that matter, the msysGit environment...

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

Make sure that 'ant' is found in the PATH

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

We actually need a JDK, not only a JRE

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

TEMP: Disable 3rdparty cloning

We aren't allowed to distribute the files under 3rdparty due to restrictions by the manufacturers. However, for this build environment to be public, this script must be able to run successfully for everyone; comment out the lines requiring SSH access.

Fixes to openspim.sh

Restored add-desktop-shortcut, had it return 0 on error so the startup
doesn't cancel, had openspim create the JDK directory... etc.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>