Completely reworked device manager
The new model basically builds an object-oriented layer on top of MM's
JNI. It's fast enough for us, but there are a few weird kinks to work out
yet. Especially for the Picard motors, which have a stepping precision of
1.5um, not everything works like we might expect (in particular, the
rotational axis calibration is now broken -- or has it been for a while?).

The biggest thing left to do is rewrite the motor sliders to use motor
steps behind the scenes: the labels will read in um (or degrees, for the
twister), but moving the motors by 1 or 10 um (i.e. by clicking the
buttons with or without shift) will move by however many *steps* bring the
motor closest to that delta.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>