Fix localization oversights
In particular, make sure that we parse and stringify doubles in a
locale-independent manner. I say that, of course, but there's no chance I
got all of them -- there may be more issues with this in the future, to be
appended as fixups to this commit.

I did, however, test the most obvious features of the plugin in both en-US
and de-DE locales and noticed no errors.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>