SPIMAcquisition: Add 'Home 4-D Stage' to plugin
So now I finally explain why all of this is necessary: on our new system,
we've found that once the stages lose power, they forget their location.
So thenceforth they report '0' to the computer, although the arm is
already quite extended; we can't retract it with a negative position.

This can be fixed by making use of the Hall-effect sensor in the
PicardStage, or more generally, by having a well-defined 'home' location
to send the stages to. This button invokes it.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>