New build scripts
With the paradigm for building changing away from compiling MM ourselves
(phew), I've started work on these new build scripts.

build-do-setup.sh is the successor to the transient setup.sh; this fetches
an MM nightly build installer and installs it.

build-mm-prereqs.sh compiles MMCore (C++ and Java) and mmstudio, in case
we make changes to these that we want to keep around. mmstudio now has a
(largely nonstandard) pom.xml, letting it be built with maven.

build-spimacq.sh builds SPIMAcquisition's Java code and deploys the
completed jar.

build-fragments.sh is a function library; pay no attention to the script
behind the curtain.

build-new.sh is the main script; it uses the others to compile a
(hopefully) working version of MicroManager + Fiji + SPIMAcquisition.

Our autostart script ceased working a while ago, with a revamp of the
plugins system in Micro-Manager. The visibility thereof is much more
carefully managed now, so the script takes an alternative approach: It
waits until the MM menu is established, then runs doClick on the
appropriate JMenuItem.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>