Redundant Asynchronous Output
WIP: Doubly asynchronous output

Unfortunately, async output is not always enough. In particular, overnight
time-lapses have apparently run into troubles, according to Pete. (And
worse, this was on a 64-bit computer with 24GB RAM.

So I'm changing strategy. Rather than limit the image queue on
construction, I'll monitor the memory available according to IJ and decide
whether to (a) enqueue a full image, (b) write the image to a temporary
file and enqueue that path, or (c) block and wait for enough space to do
(b).

Obviously, (c) is the bad choice. In the past we've avoided that using
only option (a), then falling back to (c) if there's a problem.

Option (b) sounds expensive, but it's actually not too bad -- IJ is very
quick about writing out a TIFF. It's slower than storing in memory, but it
may still buy enough time to keep acquiring images.

In general, these changes should improve the async's ability to cope with
long-term acquisitions. The thresholds are raised (from a hard line of
0.5*available memory, to 0.6*max for straight enqueue and 0.8*max for
ram-disk-ram enqueues), and using the current available memory should
allow GC to help more.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>

Doubly-asynchronous output major rework

(Need to squash this into the other commit and move that one forward.)

I changed the asynchronous paradigm pretty signficantly. Rather than
explaining the changes, I'll just explain the new handler:

There are three tweakable parameters: The memory quota, the disk quota,
and the GC aggressiveness. The handler stores slices in memory until
exceeding the memory quota, at which point it starts to write single-slice
tiffs to a temporary directory in the output directory. This continues
until the disk quota is met, at which point it starts desperately
searching for options.

While running, if the number of freed slices : total memory slices (queued
+ freed) exceeds the GC aggresiveness threshold, the handler forcibly
cycles the garbage collector twice and resets the freed slice count. This
isn't exact by any means -- GC cycles in between forced runs won't reset
the freed slice count (TODO: phantom reference queue?), and we're tracking
slices instead of bytes, but it's effective at holding off the
disk-writing. However, the GC won't be run if the handler is mid-stack
(slowed), to prevent interfering with the imaging.

The inquisitive can try playing with the settings to find the optimum
balance. I recommend a memory quota around 80% (higher than 85% triggers
runtime garbage collections, while higher than 90% risks OoM errors), a
disk quota around 90% (this can be pushed pretty high, though 100% will
cause 'disk full' errors), and a GC aggressiveness of around 50%.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>