Better integration of continuous acquisition
Previously, continuous acquisition wasn't possible with writing to disk.
This was largely because OME-TIFF can't support ragged/poorly-structured
data, which continuous acquisition rather implies: because thread
scheduling is nondeterministic, image capture rate varies, leading to
differently-sized Z stacks for the same span of Z coordinates.

A previous commit enabled the OME-TIFF handler to output such data (at the
expense of the XML structure not appropriately reflecting the data
itself), so this capability has been restored. With a few other tweaks,
continuous acquisition works with the OME-TIFF handler.

There are still serious drawbacks -- the framerate, for instance, is
apalling. It may be that our motors are too fast (over a 1500-um span, we
get ~50 images). As I mentioned in the OME-TIFF commit message, the data
structure isn't correct (the stacks aren't considered related by the
Bio-Formats Importer). Anti-drift does not work with continuous mode yet.
These issues are to be addressed in future commits.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>