ProgrammaticAcquisitor: Add AcqStatus object
This object is meant to provide some more detailed control over the
acquisition process. It contains a status flag, updated by the
performAcquisition method, which can be checked externally to monitor the
acquisition. It also contains a pause flag that will halt the acquisition
temporarily (required by IAcquisitionEngine2010). (This feature isn't very
well-tested; the behavior may be removed at a later date.)

Note that this doesn't and won't deprecate the status callback currently
used by SPIMAcquisition.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>