SPIMAcquisition: Fix start-up exception
If one of the motors is outside the accepted range, the SPIM plugin throws
an exception and won't open. For people without the Picard 4-D Stage
program to reset their positions, this could be a significant problem.
Historically I've suggested users use the MM script panel to reset stage
positions; this is a better solution.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>