SPIMAcquisition: Improvements to GenericRotator
The generic implementation of rotational stages for the SPIM device
manager needed some improvements. Some methods were removed in favor of
reusing code from the parent class, and a wrapping implementation of
setPosition was included. Rotating the twister past the end of its range
will wrap it around (so long as the user doesn't somehow reach several
thousand degrees of rotation).

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>