SPIMAcquisition: Update Live by default; move Home
By request, Update Live View has become the default setting. For most
purposes having it on shouldn't slow the system down considerably; if it
is found to, the checkbox is still there. I also moved the Home button to
sit beneath the three linear stages, since those are the only ones it
affects -- its counterpart, the Set Zero button for the twister, sits
beneath its stage as well.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>