SPIMAcquisition: UI Underhaul
The changes won't be very visible to users; the most noticable will be the
new 'sweep' button on the position list, which inserts a continuous stack.
A few buttons have been moved around as well.

Basically, initUI was getting totally out of hand, so I broke it up into
smaller methods; this should make it more navigible (in particular within
an IDE).

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>