SPIMSetup: Future-proof Device infrastructure
The old Device factory registration relied on a quirk of MM's plugin
loading which loaded each class, causing the static initializer blocks to
fire. This approach, using a ServiceLoader to find all factories, is less
elegant and more verbose, but it won't fail if MM changes the plugin
loading behavior.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>